Enforced the usage of 16-bit numpy integers when reading sparse third order force arrays in ESKM format. This limits us to a maximum of 32.3k atoms but is desirable for significantly improving memory requirements when initially reading. For an amorphous system at 250 atoms, we initialize an array of size (6, 3*n_unit * (3 * n_replicated)**2) to store coordinates, which is 5 Gb at np.int16 but blows up to 20Gb if using python's implicit "int" type. May be worth noting in our docs that this was a value-judgement and can be optimized to fit specific user's needs, but unlikely to cause trouble for most users.

+++ kaldo/interfaces/eskm_io.py
@@ -97,14 +97,14 @@ def import_sparse_third(atoms, supercell=(1, 1, 1), filename='THIRD', third_ener
     n_replicated_atoms = n_atoms * n_replicas
     n_rows = count_rows(filename)
     array_size = min(n_rows * 3, n_atoms * 3 * (n_replicated_atoms * 3) ** 2)
-    coords = np.zeros((array_size, 6), dtype=int16)
+    coords = np.zeros((array_size, 6), dtype=np.int16)
     values = np.zeros((array_size))
     index_in_unit_cell = 0
     tenjovermoltoev = 10 * units.J / units.mol
     with open(filename) as f:
         for i, line in enumerate(f):
             l_split = re.split('\s+', line.strip())
-            coords_to_write = np.array(l_split[0:-3], dtype=int) - 1
+            coords_to_write = np.array(l_split[0:-3], dtype=np.int16) - 1
             values_to_write = np.array(l_split[-3:], dtype=float)
             #TODO: add 'if' third_energy_threshold before calculating the mask
             mask_to_write = np.abs(values_to_write) > third_energy_threshold
